No reason to keep ToolStepTest and ToolStepRunTest separate.
